pyproject.toml
[tool.poetry]
name = "snakeos-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
uvicorn = "^0.31.0"
psutil = "^5.9.8"
pydantic-settings = "^2.8.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.24.2"}
asyncpg = "^0.30.0"
loguru = "^0.7.3"
distro = {version = "^1.9.0", optional = true, platform = "linux"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = ">=0.0.18"
email-validator = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
